+++ eras/allegra/impl/src/Cardano/Ledger/Allegra/Scripts.hs
@@ -52,7 +52,6 @@ where
 import Cardano.Ledger.Allegra.Era (AllegraEra)
 import Cardano.Ledger.BaseTypes (StrictMaybe (SJust, SNothing))
 import Cardano.Ledger.Binary (
-  Annotator (..),
   DecCBOR (decCBOR),
   EncCBOR (encCBOR),
   ToCBOR (..),
@@ -66,7 +65,6 @@ import Cardano.Ledger.Binary.Coders (
   encode,
   (!>),
   (<!),
-  (<*!),
  )
 import Cardano.Ledger.Core
 import Cardano.Ledger.MemoBytes (
@@ -189,20 +187,6 @@ instance Era era => EncCBOR (TimelockRaw era) where
       TimelockTimeStart m -> Sum TimelockTimeStart 4 !> To m
       TimelockTimeExpire m -> Sum TimelockTimeExpire 5 !> To m
 
--- This instance allows us to derive instance DecCBOR (Annotator (Timelock era)).
--- Since Timelock is a newtype around (Memo (Timelock era)).
-instance Era era => DecCBOR (Annotator (TimelockRaw era)) where
-  decCBOR = decode (Summands "TimelockRaw" decRaw)
-    where
-      decRaw :: Word -> Decode 'Open (Annotator (TimelockRaw era))
-      decRaw 0 = Ann (SumD TimelockSignature <! From)
-      decRaw 1 = Ann (SumD TimelockAllOf) <*! D (sequence <$> decCBOR)
-      decRaw 2 = Ann (SumD TimelockAnyOf) <*! D (sequence <$> decCBOR)
-      decRaw 3 = Ann (SumD TimelockMOf) <*! Ann From <*! D (sequence <$> decCBOR)
-      decRaw 4 = Ann (SumD TimelockTimeStart <! From)
-      decRaw 5 = Ann (SumD TimelockTimeExpire <! From)
-      decRaw n = Invalid n
-
 instance Era era => DecCBOR (TimelockRaw era) where
   decCBOR = decode $ Summands "TimelockRaw" $ \case
     0 -> SumD TimelockSignature <! From
@@ -246,9 +230,6 @@ deriving instance Show (Timelock era)
 
 instance EqRaw (Timelock era) where
   eqRaw = eqTimelockRaw
-
-instance Era era => DecCBOR (Annotator (Timelock era)) where
-  decCBOR = fmap MkTimelock <$> decCBOR
 
 -- | Since Timelock scripts are a strictly backwards compatible extension of
 -- MultiSig scripts, we can use the same 'scriptPrefixTag' tag here as we did