+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/TxWits.hs
@@ -61,6 +61,9 @@ module Cardano.Ledger.Alonzo.TxWits (
   nullDats,
   alonzoEqTxWitsRaw,
   emptyTxWitsRaw,
+  addScriptsTxWitsRaw,
+  decodeAlonzoPlutusScript,
+  asHashedScriptPair,
 )
 where
 
@@ -73,7 +76,6 @@ import Cardano.Ledger.Alonzo.Scripts (
   toPlutusSLanguage,
  )
 import Cardano.Ledger.Binary (
-  Annotator,
   DecCBOR (..),
   DecCBORGroup (..),
   Decoder,
@@ -103,7 +105,6 @@ import Cardano.Ledger.Core
 import Cardano.Ledger.Keys (BootstrapWitness, WitVKey)
 import Cardano.Ledger.MemoBytes (
   EqRaw (..),
-  Mem,
   MemoBytes,
   Memoized (..),
   eqRawType,
@@ -123,7 +124,6 @@ import Cardano.Ledger.Plutus.Language (
  )
 import Cardano.Ledger.Shelley.TxWits (
   ShelleyTxWits (..),
-  mapTraverseableDecoderA,
   shelleyEqTxWitsRaw,
  )
 import Control.DeepSeq (NFData)
@@ -346,18 +346,6 @@ nullDats :: TxDats era -> Bool
 nullDats (getMemoRawType -> TxDatsRaw m) = Map.null m
 {-# DEPRECATED nullDats "In favor of `unTxDatsL`" #-}
 
-instance Era era => DecCBOR (Annotator (TxDatsRaw era)) where
-  decCBOR =
-    ifDecoderVersionAtLeast
-      (natVersion @9)
-      ( allowTag setTag
-          >> mapTraverseableDecoderA
-            (decodeNonEmptyList decCBOR)
-            (TxDatsRaw . Map.fromElems hashData . NE.toList)
-      )
-      (mapTraverseableDecoderA (decodeList decCBOR) (TxDatsRaw . Map.fromElems hashData))
-  {-# INLINE decCBOR #-}
-
 instance Era era => DecCBOR (TxDatsRaw era) where
   decCBOR =
     ifDecoderVersionAtLeast
@@ -389,11 +377,6 @@ instance Era era => Monoid (TxDats era) where
 
 -- | Encodes memoized bytes created upon construction.
 instance Era era => EncCBOR (TxDats era)
-
-deriving via
-  Mem (TxDatsRaw era)
-  instance
-    Era era => DecCBOR (Annotator (TxDats era))
 
 -- | Upgrade 'TxDats' from one era to another. The underlying data structure
 -- will remain identical, but the memoised serialisation may change to reflect
@@ -596,51 +579,6 @@ instance AlonzoEraScript era => EncCBOR (AlonzoTxWitsRaw era) where
                 ++ show (plutusLanguage (Proxy @l))
           Just plutusScripts -> plutusScripts
 
-instance AlonzoEraScript era => DecCBOR (Annotator (RedeemersRaw era)) where
-  decCBOR = do
-    ifDecoderVersionAtLeast
-      (natVersion @9)
-      ( peekTokenType >>= \case
-          TypeMapLenIndef -> decodeMapRedeemers
-          TypeMapLen -> decodeMapRedeemers
-          _ -> decodeListRedeemers
-      )
-      ( mapTraverseableDecoderA
-          (decodeList decodeAnnElement)
-          (RedeemersRaw . Map.fromList)
-      )
-    where
-      decodeRedeemersWith nonEmptyDecoder =
-        mapTraverseableDecoderA
-          nonEmptyDecoder
-          (RedeemersRaw . Map.fromList . NE.toList)
-      decodeMapRedeemers = decodeRedeemersWith $ do
-        (_, xs) <- decodeListLikeWithCount decodeMapLenOrIndef (:) $ \_ -> do
-          ptr <- decCBOR
-          (annData, exUnits) <- decCBOR
-          pure $ (\d -> (ptr, (d, exUnits))) <$> annData
-        case NE.nonEmpty xs of
-          Nothing -> fail "Expected redeemers map to be non-empty"
-          Just neList -> pure $ NE.reverse neList
-      decodeListRedeemers =
-        decodeRedeemersWith (decodeNonEmptyList decodeAnnElement)
-      decodeAnnElement ::
-        forall s. Decoder s (Annotator (PlutusPurpose AsIx era, (Data era, ExUnits)))
-      decodeAnnElement = do
-        (rdmrPtr, dat, ex) <- decodeElement
-        let f x y z = (x, (y, z))
-        pure $ f rdmrPtr <$> dat <*> pure ex
-      {-# INLINE decodeAnnElement #-}
-      decodeElement ::
-        forall s. Decoder s (PlutusPurpose AsIx era, Annotator (Data era), ExUnits)
-      decodeElement = do
-        decodeRecordNamed
-          "Redeemer"
-          (\(rdmrPtr, _, _) -> fromIntegral (listLen rdmrPtr) + 2)
-          $ (,,) <$> decCBORGroup <*> decCBOR <*> decCBOR
-      {-# INLINE decodeElement #-}
-  {-# INLINE decCBOR #-}
-
 instance AlonzoEraScript era => DecCBOR (RedeemersRaw era) where
   decCBOR =
     ifDecoderVersionAtLeast
@@ -678,85 +616,7 @@ instance AlonzoEraScript era => DecCBOR (RedeemersRaw era) where
 -- | Encodes memoized bytes created upon construction.
 instance AlonzoEraScript era => EncCBOR (Redeemers era)
 
-deriving via
-  Mem (RedeemersRaw era)
-  instance
-    AlonzoEraScript era => DecCBOR (Annotator (Redeemers era))
-
-deriving via
-  Mem (AlonzoTxWitsRaw era)
-  instance
-    AlonzoEraScript era => DecCBOR (Annotator (AlonzoTxWits era))
-
-instance
-  AlonzoEraScript era =>
-  DecCBOR (Annotator (AlonzoTxWitsRaw era))
-  where
-  decCBOR =
-    decode $
-      SparseKeyed
-        "AlonzoTxWits"
-        (pure emptyTxWitsRaw)
-        txWitnessField
-        []
-    where
-      txWitnessField :: Word -> Field (Annotator (AlonzoTxWitsRaw era))
-      txWitnessField 0 =
-        fieldAA
-          (\x wits -> wits {atwrAddrTxWits = x})
-          ( D $
-              ifDecoderVersionAtLeast
-                (natVersion @9)
-                ( allowTag setTag
-                    >> mapTraverseableDecoderA (decodeNonEmptyList decCBOR) (Set.fromList . NE.toList)
-                )
-                (mapTraverseableDecoderA (decodeList decCBOR) Set.fromList)
-          )
-      txWitnessField 1 =
-        fieldAA
-          addScriptsTxWitsRaw
-          (D nativeScriptsDecoder)
-      txWitnessField 2 =
-        fieldAA
-          (\x wits -> wits {atwrBootAddrTxWits = x})
-          ( D $
-              ifDecoderVersionAtLeast
-                (natVersion @9)
-                ( allowTag setTag
-                    >> mapTraverseableDecoderA (decodeNonEmptyList decCBOR) (Set.fromList . NE.toList)
-                )
-                (mapTraverseableDecoderA (decodeList decCBOR) Set.fromList)
-          )
-      txWitnessField 3 = fieldA addScriptsTxWitsRaw (decodePlutus SPlutusV1)
-      txWitnessField 4 =
-        fieldAA
-          (\x wits -> wits {atwrDatsTxWits = x})
-          From
-      txWitnessField 5 = fieldAA (\x wits -> wits {atwrRdmrsTxWits = x}) From
-      txWitnessField 6 = fieldA addScriptsTxWitsRaw (decodePlutus SPlutusV2)
-      txWitnessField 7 = fieldA addScriptsTxWitsRaw (decodePlutus SPlutusV3)
-      txWitnessField n = invalidField n
-      {-# INLINE txWitnessField #-}
-
-      nativeScriptsDecoder :: Decoder s (Annotator (Map ScriptHash (Script era)))
-      nativeScriptsDecoder =
-        ifDecoderVersionAtLeast
-          (natVersion @9)
-          ( allowTag setTag
-              >> mapTraverseableDecoderA (decodeNonEmptyList pairDecoder) (Map.fromList . NE.toList)
-          )
-          (mapTraverseableDecoderA (decodeList pairDecoder) Map.fromList)
-        where
-          pairDecoder :: Decoder s (Annotator (ScriptHash, Script era))
-          pairDecoder = fmap (asHashedPair . fromNativeScript) <$> decCBOR
-  {-# INLINE decCBOR #-}
-
-instance
-  ( AlonzoEraScript era
-  , DecCBOR (NativeScript era)
-  ) =>
-  DecCBOR (AlonzoTxWitsRaw era)
-  where
+instance AlonzoEraScript era => DecCBOR (AlonzoTxWitsRaw era) where
   decCBOR =
     decode $
       SparseKeyed
@@ -789,11 +649,11 @@ instance
                 )
                 (Set.fromList <$> decodeList decCBOR)
           )
-      txWitnessField 3 = field addScriptsTxWitsRaw (decodePlutus SPlutusV1)
+      txWitnessField 3 = field addScriptsTxWitsRaw (decodeAlonzoPlutusScript SPlutusV1)
       txWitnessField 4 = field (\x wits -> wits {atwrDatsTxWits = x}) From
       txWitnessField 5 = field (\x wits -> wits {atwrRdmrsTxWits = x}) From
-      txWitnessField 6 = field addScriptsTxWitsRaw (decodePlutus SPlutusV2)
-      txWitnessField 7 = field addScriptsTxWitsRaw (decodePlutus SPlutusV3)
+      txWitnessField 6 = field addScriptsTxWitsRaw (decodeAlonzoPlutusScript SPlutusV2)
+      txWitnessField 7 = field addScriptsTxWitsRaw (decodeAlonzoPlutusScript SPlutusV3)
       txWitnessField n = invalidField n
 
       nativeScriptsDecoder :: Decoder s (Map ScriptHash (Script era))
@@ -806,13 +666,9 @@ instance
           (Map.fromList <$> decodeList pairDecoder)
         where
           pairDecoder :: Decoder s (ScriptHash, Script era)
-          pairDecoder = asHashedPair @era . fromNativeScript <$> decCBOR
+          pairDecoder = asHashedScriptPair @era . fromNativeScript <$> decCBOR
 
-deriving newtype instance
-  ( AlonzoEraScript era
-  , DecCBOR (NativeScript era)
-  ) =>
-  DecCBOR (AlonzoTxWits era)
+deriving newtype instance AlonzoEraScript era => DecCBOR (AlonzoTxWits era)
 
 addScriptsTxWitsRaw ::
   Map ScriptHash (Script era) ->
@@ -824,17 +680,17 @@ addScriptsTxWitsRaw scriptWitnesses txWits =
     }
 {-# INLINE addScriptsTxWitsRaw #-}
 
-decodePlutus ::
+decodeAlonzoPlutusScript ::
   (AlonzoEraScript era, PlutusLanguage l) =>
   SLanguage l ->
   Decode ('Closed 'Dense) (Map ScriptHash (Script era))
-decodePlutus slang =
+decodeAlonzoPlutusScript slang =
   D $
     ifDecoderVersionAtLeast
       (natVersion @9)
       (scriptDecoderV9 (fromPlutusScript <$> decodePlutusScript slang))
       (scriptDecoder (fromPlutusScript <$> decodePlutusScript slang))
-{-# INLINE decodePlutus #-}
+{-# INLINE decodeAlonzoPlutusScript #-}
 
 scriptDecoderV9 ::
   EraScript era =>
@@ -843,7 +699,7 @@ scriptDecoderV9 ::
 scriptDecoderV9 decodeScript = do
   allowTag setTag
   scriptMap <- decodeMapLikeEnforceNoDuplicates decodeListLenOrIndef $ do
-    asHashedPair <$> decodeScript
+    asHashedScriptPair <$> decodeScript
   when (Map.null scriptMap) $ fail "Empty list of scripts is not allowed"
   pure scriptMap
 {-# INLINE scriptDecoderV9 #-}
@@ -855,14 +711,14 @@ scriptDecoder ::
 scriptDecoder decodeScript =
   fmap Map.fromList $
     decodeList $
-      asHashedPair <$> decodeScript
+      asHashedScriptPair <$> decodeScript
 {-# INLINE scriptDecoder #-}
 
-asHashedPair :: forall era. EraScript era => Script era -> (ScriptHash, Script era)
-asHashedPair script =
+asHashedScriptPair :: forall era. EraScript era => Script era -> (ScriptHash, Script era)
+asHashedScriptPair script =
   let !scriptHash = hashScript @era script
    in (scriptHash, script)
-{-# INLINE asHashedPair #-}
+{-# INLINE asHashedScriptPair #-}
 
 alonzoEqTxWitsRaw :: AlonzoEraTxWits era => TxWits era -> TxWits era -> Bool
 alonzoEqTxWitsRaw txWits1 txWits2 =