+++ eras/mary/impl/src/Cardano/Ledger/Mary/TxBody/Internal.hs
@@ -48,7 +48,7 @@ where
 
 import Cardano.Ledger.Allegra.Core
 import Cardano.Ledger.Allegra.TxBody
-import Cardano.Ledger.Binary (Annotator, DecCBOR (..), EncCBOR (..), ToCBOR (..))
+import Cardano.Ledger.Binary (DecCBOR (..), EncCBOR (..), ToCBOR (..))
 import Cardano.Ledger.Coin (Coin (..))
 import Cardano.Ledger.Mary.Era (MaryEra)
 import Cardano.Ledger.Mary.TxCert ()
@@ -56,7 +56,6 @@ import Cardano.Ledger.Mary.TxOut ()
 import Cardano.Ledger.Mary.Value
 import Cardano.Ledger.MemoBytes (
   EqRaw,
-  Mem,
   MemoBytes,
   MemoHashIndex,
   Memoized (RawType),
@@ -120,11 +119,6 @@ deriving newtype instance
 deriving newtype instance
   (Era era, NFData (TxOut era), NFData (TxCert era), NFData (PParamsUpdate era)) =>
   NFData (MaryTxBody era)
-
-deriving via
-  Mem (MaryTxBodyRaw era)
-  instance
-    MaryEraTxBody era => DecCBOR (Annotator (MaryTxBody era))
 
 deriving newtype instance MaryEraTxBody era => DecCBOR (MaryTxBody era)