+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/TxWits.hs
@@ -19,7 +19,6 @@
 {-# OPTIONS_GHC -Wno-orphans #-}
 
 module Cardano.Ledger.Shelley.TxWits (
-  decodeWits,
   ShelleyTxWits (
     MkShelleyTxWits,
     ShelleyTxWits,
@@ -33,14 +32,11 @@ module Cardano.Ledger.Shelley.TxWits (
   bootAddrShelleyTxWitsL,
   addrWits',
   shelleyEqTxWitsRaw,
-  mapTraverseableDecoderA,
 )
 where
 
 import Cardano.Ledger.Binary (
-  Annotator,
   DecCBOR (decCBOR),
-  Decoder,
   EncCBOR (encCBOR),
   decodeList,
  )
@@ -57,7 +53,6 @@ import Cardano.Ledger.Hashes (SafeToHash (..))
 import Cardano.Ledger.Keys (BootstrapWitness, KeyRole (Witness), WitVKey (..))
 import Cardano.Ledger.MemoBytes (
   EqRaw (..),
-  Mem,
   MemoBytes,
   Memoized (..),
   decodeMemoized,
@@ -75,7 +70,6 @@ import qualified Data.Map.Strict as Map
 import qualified Data.MapExtras as Map (fromElems)
 import Data.Set (Set)
 import qualified Data.Set as Set
-import Data.Void
 import GHC.Generics (Generic)
 import GHC.Records ()
 import Lens.Micro (Lens', (^.))
@@ -127,7 +121,7 @@ instance
 
 instance EraScript era => NoThunks (ShelleyTxWits era)
 
-instance (EraScript era, DecCBOR (Script era)) => DecCBOR (ShelleyTxWits era) where
+instance EraScript era => DecCBOR (ShelleyTxWits era) where
   decCBOR = MkShelleyTxWits <$> decodeMemoized decCBOR
 
 -- =======================================================
@@ -222,15 +216,7 @@ shelleyEqTxWitsRaw txWits1 txWits2 =
     && liftEq eqRaw (txWits1 ^. scriptTxWitsL) (txWits2 ^. scriptTxWitsL)
     && txWits1 ^. bootAddrTxWitsL == txWits2 ^. bootAddrTxWitsL
 
-instance EraScript era => DecCBOR (Annotator (ShelleyTxWitsRaw era)) where
-  decCBOR = decodeWits
-
-deriving via
-  Mem (ShelleyTxWitsRaw era)
-  instance
-    EraScript era => DecCBOR (Annotator (ShelleyTxWits era))
-
-instance (EraScript era, DecCBOR (Script era)) => DecCBOR (ShelleyTxWitsRaw era) where
+instance EraScript era => DecCBOR (ShelleyTxWitsRaw era) where
   decCBOR =
     decode $
       SparseKeyed
@@ -247,42 +233,3 @@ instance (EraScript era, DecCBOR (Script era)) => DecCBOR (ShelleyTxWitsRaw era)
           (D $ Map.fromElems (hashScript @era) <$> decodeList decCBOR)
       witField 2 = field (\x wits -> wits {stwrBootAddrTxWits = x}) From
       witField n = invalidField n
-
-decodeWits ::
-  forall era s.
-  EraScript era =>
-  Decoder s (Annotator (ShelleyTxWitsRaw era))
-decodeWits =
-  decode $
-    SparseKeyed
-      "ShelleyTxWitsRaw"
-      (pure emptyWitnessSet)
-      witField
-      []
-  where
-    emptyWitnessSet = ShelleyTxWitsRaw mempty mempty mempty
-    witField :: Word -> Field (Annotator (ShelleyTxWitsRaw era))
-    witField 0 =
-      fieldAA
-        (\x wits -> wits {stwrAddrTxWits = x})
-        (D $ mapTraverseableDecoderA (decodeList decCBOR) Set.fromList)
-    witField 1 =
-      fieldAA
-        (\x wits -> wits {stwrScriptTxWits = x})
-        ( D $
-            mapTraverseableDecoderA
-              (decodeList decCBOR)
-              (Map.fromElems (hashScript @era))
-        )
-    witField 2 =
-      fieldAA
-        (\x wits -> wits {stwrBootAddrTxWits = x})
-        (D $ mapTraverseableDecoderA (decodeList decCBOR) Set.fromList)
-    witField n = fieldAA (\(_ :: Void) wits -> wits) (Invalid n)
-
-mapTraverseableDecoderA ::
-  Traversable f =>
-  Decoder s (f (Annotator a)) ->
-  (f a -> m b) ->
-  Decoder s (Annotator (m b))
-mapTraverseableDecoderA decList transformList = fmap transformList . sequence <$> decList